feat: get fintecture access token

+++ src/main/java/app/bpartners/api/model/Transaction.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 @EqualsAndHashCode
 public class Transaction {
   private String id;
-  private Double amount;
+  private int amount;
   private String currency;
   private String label;
   private String reference;
